+++ CHANGELOG.md
@@ -13,7 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `:` symbols. E.g. `|:--|:--:|--:|` will set left, center and right alignment for 3 columns respectively.
 - \[Markdown] Support for table width specification in percent. You can specify table width using `{width=75%}`
   attribute under the table (percent sign is optional).
-- \[Asciidoc] Support for text alignment with `[.align-left]`, `[.align-center]` and `[.align-right]` classes.
+- \[Asciidoc] Support for text alignment with `[.text-left]`, `[.text-center]` and `[.text-right]` classes.
 - \[Asciidoc] Support for the following table features:
   - Table cell background colors (same as in Markdown)
   - Column/column text alignment (both vertical and horizontal)